// src/app/factory-planner-api/model/crafting-site-node.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FactorySiteIO } from './factory-site-io';
import { FactoryNode } from './factory-node';
import { CraftingMachineEntity } from './crafting-machine-entity';
import { ItemDescriptorDto } from './item-descriptor-dto';
import { RecipeProducingDto } from './recipe-producing-dto';


export interface CraftingSiteNode extends FactoryNode { 
    automaton: CraftingMachineEntity;
    recipe: RecipeProducingDto;
    requiredMachines: number;
    produces: Array<FactorySiteIO>;
    manufacturingDuration: number;
    cyclePerMinute: number;
    targetOutputPerMinute: number;
}
export namespace CraftingSiteNode {
}